1,修改图文帖界面； 2，修改图文帖收藏，举报 3，修改送礼物界面； 4，修改剩余金币字段

+++ MidnightArtifact/app/src/main/java/com/atman/wysq/model/request/AddCommentModel.java
@@ -8,10 +8,10 @@ package com.atman.wysq.model.request;
  * 电话 18578909061
  */
 public class AddCommentModel {
-    private int blog_id;
+    private long blog_id;
     private String content;
 
-    public AddCommentModel(int blog_id, String content) {
+    public AddCommentModel(long blog_id, String content) {
         this.blog_id = blog_id;
         this.content = content;
     }
